Replace forked client with net/smtp
