+++ auth.go
@@ -3,6 +3,7 @@ package main
 import (
 	"bufio"
 	"errors"
+	"net/smtp"
 	"os"
 	"strings"
 
@@ -97,4 +98,31 @@ func AuthCheckPassword(username string, secret string) error {
 		return nil
 	}
 	return errors.New("password invalid")
+}
+
+// LOGIN authentication
+type loginAuth struct {
+	username, password string
+}
+
+func LoginAuth(username, password string) smtp.Auth {
+	return &loginAuth{username, password}
+}
+
+func (a *loginAuth) Start(_ *smtp.ServerInfo) (string, []byte, error) {
+	return "LOGIN", []byte{}, nil
+}
+
+func (a *loginAuth) Next(fromServer []byte, more bool) ([]byte, error) {
+	if more {
+		switch string(fromServer) {
+		case "Username:":
+			return []byte(a.username), nil
+		case "Password:":
+			return []byte(a.password), nil
+		default:
+			return nil, errors.New("unknown fromServer")
+		}
+	}
+	return nil, nil
 }
